+++ package.json
@@ -19,6 +19,7 @@
     "clsx": "^2.1.1",
     "dompurify": "^3.2.5",
     "i": "^0.3.7",
+    "lucide-react": "^0.503.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hot-toast": "^2.5.2",
